fix: boss health staleness, float damage fields, upsert on re-parse
- Parser: record health=0 in health_samples on boss entity deletion so
  build_health_window carry-forward correctly shows dead objectives as 0
  rather than their last non-zero damage sample
- Backend: widen DamageRecord.pre_damage and damage_absorbed from int to
  float -- Deadlock applies fractional damage internally; parser emits f32
- Backend: replace INSERT + pgcode conflict check with pg_insert upsert;
  the old pgcode check silently broke with asyncpg (.sqlstate not .pgcode)
  causing a 500 on any re-parse; upsert also ensures re-parses refresh
  stored data rather than keeping stale cached rows

+++ backend/app/domain/player.py
@@ -37,12 +37,12 @@ class PlayerData(SQLModel):
 
 class DamageRecord(SQLModel):
     damage: Optional[int] = None
-    pre_damage: Optional[int] = None
+    pre_damage: Optional[float] = None
     type: Optional[int] = None
     citadel_type: Optional[int] = None
     entindex_inflictor: Optional[int] = None
     entindex_ability: Optional[int] = None
-    damage_absorbed: Optional[int] = None
+    damage_absorbed: Optional[float] = None
     victim_health_max: Optional[int] = None
     victim_health_new: Optional[int] = None
     flags: Optional[int] = None
